# src/api/openapi.yaml
openapi: 3.0.3
info:
  title: Cloud Storage Transfer Service API
  description: REST API for transferring files between Azure Blob Storage and local storage
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.example.com
    description: Production server (Cloud Run)

paths:
  /api/v1/transfer/download:
    post:
      summary: Download files from Azure Blob Storage
      description: Initiates a download transfer job for files from Azure Blob Storage to local storage
      operationId: downloadFiles
      tags:
        - Transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                source_url:
                  type: string
                  description: Azure Blob Storage URL (optional if using env config)
                  example: "https://account.blob.core.windows.net/container"
                sas_token:
                  type: string
                  description: SAS token for authentication (optional if using env config)
                  example: "?sv=2021-06-08&ss=..."
                prefix:
                  type: string
                  description: Filter files by prefix
                  example: "documents/"
                max_files:
                  type: integer
                  description: Maximum number of files to download
                  example: 100
                destination_path:
                  type: string
                  description: Destination directory path
                  example: "./downloads"
                overwrite:
                  type: boolean
                  description: Whether to overwrite existing files
                  default: false
              example:
                prefix: "reports/"
                max_files: 50
                overwrite: true
      responses:
        '202':
          description: Transfer job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferJobResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/transfer/upload:
    post:
      summary: Upload a file to Azure Blob Storage
      description: Uploads a single file to Azure Blob Storage
      operationId: uploadFile
      tags:
        - Transfer
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                destination_path:
                  type: string
                  description: Destination path in blob storage
                  example: "uploads/myfile.txt"
                date_folder:
                  type: boolean
                  description: Create and upload to date folder (YYYYMMDD)
                  default: false
                overwrite:
                  type: boolean
                  description: Overwrite if file exists
                  default: false
                container_url:
                  type: string
                  description: Azure container URL (optional if using env config)
                sas_token:
                  type: string
                  description: SAS token (optional if using env config)
      responses:
        '200':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadResponse'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Conflict - file exists and overwrite is false
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/transfer/{job_id}:
    get:
      summary: Get transfer job status
      description: Retrieves the status of a transfer job
      operationId: getTransferStatus
      tags:
        - Transfer
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
          description: Transfer job ID
      responses:
        '200':
          description: Transfer job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferJobStatus'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /health:
    get:
      summary: Health check endpoint
      description: Check if the API is running
      operationId: healthCheck
      tags:
        - System
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "healthy"
                  timestamp:
                    type: string
                    format: date-time
                  version:
                    type: string
                    example: "1.0.0"

components:
  schemas:
    TransferJobResponse:
      type: object
      properties:
        job_id:
          type: string
          description: Unique job identifier
          example: "job_20240908_123456"
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
          example: "in_progress"
        message:
          type: string
          example: "Transfer job initiated"
        created_at:
          type: string
          format: date-time
        destination_path:
          type: string
          example: "./downloads/20240908_123456"

    TransferJobStatus:
      type: object
      properties:
        job_id:
          type: string
          example: "job_20240908_123456"
        status:
          type: string
          enum: [pending, in_progress, completed, failed]
        progress:
          type: object
          properties:
            total_files:
              type: integer
              example: 100
            completed_files:
              type: integer
              example: 45
            failed_files:
              type: integer
              example: 2
            skipped_files:
              type: integer
              example: 5
            total_bytes:
              type: integer
              format: int64
              example: 1048576000
            transferred_bytes:
              type: integer
              format: int64
              example: 524288000
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
        duration_seconds:
          type: number
          format: float
          example: 45.6
        errors:
          type: array
          items:
            type: string
          example: ["Failed to download file1.txt: Access denied"]

    UploadResponse:
      type: object
      properties:
        status:
          type: string
          enum: [success, failed]
          example: "success"
        message:
          type: string
          example: "File uploaded successfully"
        blob_name:
          type: string
          example: "uploads/myfile.txt"
        blob_url:
          type: string
          example: "https://account.blob.core.windows.net/container/uploads/myfile.txt"
        size:
          type: integer
          format: int64
          example: 102400
        md5_hash:
          type: string
          example: "5d41402abc4b2a76b9719d911017c592"
        uploaded_at:
          type: string
          format: date-time

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Invalid request"
        detail:
          type: string
          example: "The SAS token provided is invalid or expired"
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          example: "req_123456"

  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication (optional)

security:
  - ApiKeyAuth: []
  - {}